# add SingleFieldWriter for progressive (== faster) serialization