@BurntSushi 2018-08-23
This PR is on crates.io in `csv 1.0.1`.